Add first few tests

+++ controller.py
@@ -13,12 +13,12 @@ api_version = "v1"
 
 class CredStashController:
     def __init__(
-        self,
-        access_key_id,
-        secret_access_key,
-        default_region,
-        default_table,
-        namespaces,
+            self,
+            access_key_id,
+            secret_access_key,
+            default_region,
+            default_table,
+            namespaces,
     ):
 
         self.access_key_id = access_key_id
@@ -55,11 +55,13 @@ class CredStashController:
                     "namespace": namespace,
                     "annotations": {"credstash-fully-managed": "true"},
                 }
-                secret_obj = client.V1Secret(api_version, {}, "Secret", metadata)
+                secret_obj = client.V1Secret(api_version, {}, "Secret",
+                                             metadata)
             if (
-                new
-                or secret_obj.metadata.annotations.get("credstash-fully-managed", None)
-                == "true"
+                    new
+                    or secret_obj.metadata.annotations.get(
+                "credstash-fully-managed", None)
+                    == "true"
             ):
                 secret_obj.data = {}
             for secret_to_process in keys:
@@ -76,7 +78,8 @@ class CredStashController:
                         aws_secret_access_key=self.secret_access_key,
                         region=self.default_region,
                     )
-                    secret_obj.data[secret_to_process["key"]] = base64.b64encode(
+                    secret_obj.data[
+                        secret_to_process["key"]] = base64.b64encode(
                         raw_secret.encode()
                     ).decode()
                 except ClientError:
@@ -86,20 +89,23 @@ class CredStashController:
                 except credstash.ItemNotFound:
                     print(
                         "ERROR: {} version {} not found, bailing out!".format(
-                            secret_to_process["from"], secret_to_process["version"]
+                            secret_to_process["from"],
+                            secret_to_process["version"]
                         )
                     )
                     return
                 except KeyError as e:
                     print(
-                        "{} is missing for this secret, bailing out!".format(e.args[0])
+                        "{} is missing for this secret, bailing out!".format(
+                            e.args[0])
                     )
                     return
 
             if new:
                 self.v1core.create_namespaced_secret(namespace, secret_obj)
             else:
-                self.v1core.patch_namespaced_secret(secret, namespace, secret_obj)
+                self.v1core.patch_namespaced_secret(secret, namespace,
+                                                    secret_obj)
 
     def process_event(self, event):
         print("Event received.")
@@ -110,7 +116,9 @@ class CredStashController:
             return
         namespace = obj["metadata"]["namespace"]
         if self.namespaces is not None and namespace not in self.namespaces:
-            print("Secret requested from an " "unauthorized namespace, skipping.")
+            print(
+                "Secret requested from an "
+                "unauthorized namespace, skipping.")
             return
         metadata = obj.get("metadata")
 
@@ -126,25 +134,36 @@ class CredStashController:
         print("Waiting for credstash secrets to be defined...")
         while True:
             stream = watch.Watch().stream(
-                self.crds.list_cluster_custom_object, DOMAIN, "v1", "credstashsecrets"
+                self.crds.list_cluster_custom_object, DOMAIN, "v1",
+                "credstashsecrets"
             )
             for event in stream:
                 self.process_event(event)
 
     def delete_secret(self, credstash_secret):
-        spec = credstash_secret["spec"]
+        spec = credstash_secret.get("spec", None)
+        if not spec:
+            print("Spec empty, bailing out..")
+            return
         namespace = credstash_secret["metadata"]["namespace"]
 
         for secret, keys in spec.items():
-            secret_obj = self.v1core.read_namespaced_secret(secret, namespace=namespace)
-            print("{} is managed by credstash, deleting it".format(secret))
-            if (
-                secret_obj.metadata.annotations.get("credstash-fully-managed", None)
-                == "true"
-            ):
+            secret_obj = self.v1core.read_namespaced_secret(
+                secret,
+                namespace=namespace)
+
+            if (secret_obj.metadata.annotations.get("credstash-fully-managed",
+                                                    None) == "true"
+                ):
+                print(
+                    "{} is managed by credstash, deleting it".format(secret))
                 self.v1core.delete_namespaced_secret(
                     secret, namespace, V1DeleteOptions()
                 )
+            else:
+                print(
+                    "{} is NOT managed by credstash, NOT deleting it".format(
+                        secret))
 
 
 if __name__ == "__main__":
@@ -156,15 +175,16 @@ if __name__ == "__main__":
     main_access_key_id = os.environ["CREDSTASH_AWS_ACCESS_KEY_ID"]
     main_secret_access_key = os.environ["CREDSTASH_AWS_SECRET_ACCESS_KEY"]
     main_default_region = os.environ["CREDSTASH_AWS_DEFAULT_REGION"]
-    main_default_table = os.environ.get("CREDSTASH_DEFAULT_TABLE", "credential-store")
-    namespaces = os.environ.get("namespaces", "*")
+    main_default_table = os.environ.get("CREDSTASH_DEFAULT_TABLE",
+                                        "credential-store")
+    main_namespaces = os.environ.get("namespaces", "*")
 
     credstash_controller = CredStashController(
         main_access_key_id,
         main_secret_access_key,
         main_default_region,
         main_default_table,
-        namespaces,
+        main_namespaces,
     )
 
     credstash_controller.main_loop()
